💻 Started system for creating rewards via gui
