+++ src/main/java/lootcrate/commands/subs/SubCommandLootCrateCreate.java
@@ -6,12 +6,9 @@ import lootcrate.commands.SubCommand;
 import lootcrate.enums.Message;
 import lootcrate.enums.Permission;
 import lootcrate.enums.Placeholder;
-import lootcrate.gui.frames.creation.CrateCreateFrame;
 import lootcrate.objects.Crate;
 import lootcrate.utils.CommandUtils;
-import lootcrate.utils.DebugUtils;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 import java.util.LinkedList;
 import java.util.List;
